+++ halaman/periode_sensus/tampil.php
@@ -45,43 +45,49 @@
                                 $no = 1;
                                 ?>
                                 <tbody>
-                                    <?php while ($row = $result->fetch_assoc()) : ?>
-                                        <tr>
-                                            <td class="text-center fit align-middle">
-                                                <p><?= $no++; ?></p>
-                                            </td>
-                                            <td class="text-center align-middle">
-                                                <p><?= $row['tahun']; ?></p>
-                                            </td>
-                                            <td class="text-center align-middle">
-                                                <p><?= $row['tanggal_mulai']; ?></p>
-                                            </td>
-                                            <td class="text-center align-middle">
-                                                <p><?= $row['tanggal_selesai']; ?></p>
-                                            </td>
-                                            <td class="text-center fit">
-                                                <?php if ($row['status'] == 'Menunggu') : ?>
-                                                    <span class="status-btn info-btn">Menunggu</span>
+                                    <?php if ($result->num_rows) : ?>
+                                        <?php while ($row = $result->fetch_assoc()) : ?>
+                                            <tr>
+                                                <td class="text-center fit align-middle">
+                                                    <p><?= $no++; ?></p>
+                                                </td>
+                                                <td class="text-center align-middle">
+                                                    <p><?= $row['tahun']; ?></p>
+                                                </td>
+                                                <td class="text-center align-middle">
+                                                    <p><?= $row['tanggal_mulai']; ?></p>
+                                                </td>
+                                                <td class="text-center align-middle">
+                                                    <p><?= $row['tanggal_selesai']; ?></p>
+                                                </td>
+                                                <td class="text-center">
+                                                    <?php if ($row['status'] == 'Menunggu') : ?>
+                                                        <span class="status-btn info-btn">Menunggu</span>
                                                     <?php elseif ($row['status'] == 'Berjalan') : ?>
                                                         <span class="status-btn active-btn">Berjalan</span>
-                                                        <?php elseif ($row['status'] == 'Selesai') : ?>
-                                                            <span class="status-btn success-btn">Selesai</span>
-                                                <?php endif; ?>
-                                            </td>
-                                            <td class="d-flex gap-2 fit pt-4">
-                                                <div class="action">
-                                                    <a href="?page=periode_sensus&action=ubah&id=<?= $row['id']; ?>" class="text-warning">
-                                                        <i class="lni lni-pencil"></i>
-                                                    </a>
-                                                </div>
-                                                <div class="action">
-                                                    <a onclick="return confirm('Yakin?')" href="?page=periode_sensus&action=hapus&id=<?= $row['id']; ?>" class="text-danger">
-                                                        <i class="lni lni-trash-can"></i>
-                                                    </a>
-                                                </div>
-                                            </td>
+                                                    <?php elseif ($row['status'] == 'Selesai') : ?>
+                                                        <span class="status-btn success-btn">Selesai</span>
+                                                    <?php endif; ?>
+                                                </td>
+                                                <td class="d-flex gap-2 fit pt-4">
+                                                    <div class="action">
+                                                        <a href="?page=periode_sensus&action=ubah&id=<?= $row['id']; ?>" class="text-warning">
+                                                            <i class="lni lni-pencil"></i>
+                                                        </a>
+                                                    </div>
+                                                    <div class="action">
+                                                        <a onclick="return confirm('Yakin?')" href="?page=periode_sensus&action=hapus&id=<?= $row['id']; ?>" class="text-danger">
+                                                            <i class="lni lni-trash-can"></i>
+                                                        </a>
+                                                    </div>
+                                                </td>
+                                            </tr>
+                                        <?php endwhile; ?>
+                                    <?php else : ?>
+                                        <tr>
+                                            <td class="text-center" colspan="5">Data Kosong</td>
                                         </tr>
-                                    <?php endwhile; ?>
+                                    <?php endif; ?>
                                 </tbody>
                             </table>
                         </div>
@@ -91,21 +97,4 @@
         </div>
     </div>
 </section>
-<script src="assets/js/bootstrap.bundle.min.js"></script>
-<script src="assets/js/Chart.min.js"></script>
-<script src="assets/js/apexcharts.min.js"></script>
-<script src="assets/js/dynamic-pie-chart.js"></script>
-<script src="assets/js/moment.min.js"></script>
-<script src="assets/js/fullcalendar.js"></script>
-<script src="assets/js/jvectormap.min.js"></script>
-<script src="assets/js/world-merc.js"></script>
-<script src="assets/js/polyfill.js"></script>
-<script src="assets/js/quill.min.js"></script>
-<script src="assets/js/datatable.js"></script>
-<script src="assets/js/Sortable.min.js"></script>
-<script src="assets/js/main.js"></script>
-<script>
-    const dataTable = new simpleDatatables.DataTable("#table", {
-        searchable: true,
-    });
-</script>
+<?php include_once('layout/js.php'); ?>
